Now boots on physical hardware using the iso.sh script to generate an iso image.

+++ libc/include/stdlib.h
@@ -11,7 +11,7 @@ extern "C" {
 __attribute__((__noreturn__))
 void abort(void);
 
-char* itoa(int value, char* str, int base);
+char* itoa(long long int value, char* str, int base);
 char* ftoa(double value, char* str, int precision);
 
 // Memory functions
